Revamp run-demo; add collect-logs, logs/*-wsl.log
Now `run-demo` can be executed in `--debug` or `--strace` mode, and
shuts the web server down more cleanly thanks to using an EXIT trap.

`collect-logs` automatically runs `run-demo` multiple times to capture
four permutations of the logs (`--debug` or `--strace`, with and without
`--no-ws`). The strace logs are heavily edited to show the activity near
the point of the successful page load (`strace-no-websocket-wsl.log`) or
the crash (`strace-wsl.log`).

The previous `logs/*-ubuntu1704.txt` logs have been removed and will be
regenerated in the next commit.